requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
Jinja2
python-dotenv
